+++ src/user/user.controller.ts
@@ -1,7 +1,6 @@
 import {
   Controller,
   Get,
-  Post,
   Body,
   Patch,
   Param,
@@ -10,63 +9,68 @@ import {
   Req,
 } from '@nestjs/common';
 import { UserService } from './user.service';
-import { CreateUserDto } from './dto/create-user.dto';
 import { UpdateUserDto } from './dto/update-user.dto';
 import { AuthGuard } from 'src/auth/auth.guard';
 import { ApiOperation } from '@nestjs/swagger';
 
 @Controller('user')
 export class UserController {
-  // constructor(private readonly userService: UserService) {}
-  // @Post()
-  // @ApiOperation({
-  //   summary: 'Crée un utilisateur',
-  // })
-  // create(@Body() createUserDto: CreateUserDto) {
-  //   return this.userService.create(createUserDto);
-  // }
-  // @Get()
-  // @ApiOperation({
-  //   summary: 'Récupère tous les utilisateurs',
-  // })
-  // findAll() {
-  //   return this.userService.findAll();
-  // }
-  // @Get(':auth_id')
-  // @ApiOperation({
-  //   summary: 'Récupère un utilisateur',
-  // })
-  // findOne(@Param('auth_id') auth_id: string) {
-  //   return this.userService.findOne(auth_id);
-  // }
-  // @UseGuards(AuthGuard)
-  // @Patch()
-  // @ApiOperation({
-  //   summary:
-  //     "Met à jour un utilisateur, seul l'utilisateur authentifié peut le faire",
-  // })
-  // update(@Req() request: Request, @Body() updateUserDto: UpdateUserDto) {
-  //   const auth_id = request['user'].id;
-  //   return this.userService.update(auth_id, updateUserDto);
-  // }
-  // @UseGuards(AuthGuard)
-  // @Patch('/deactivate')
-  // @ApiOperation({
-  //   summary:
-  //     "Désactive un utilisateur, seul l'utilisateur authentifié peut le faire",
-  // })
-  // deactivate(@Req() request: Request) {
-  //   const auth_id = request['user'].id;
-  //   return this.userService.deactivate(auth_id);
-  // }
-  // @UseGuards(AuthGuard)
-  // @Delete(':auth_id')
-  // @ApiOperation({
-  //   summary:
-  //     "Supprime un utilisateur, seul l'utilisateur authentifié peut le faire",
-  // })
-  // remove(@Req() request: Request) {
-  //   const auth_id = request['user'].id;
-  //   return this.userService.remove(auth_id);
-  // }
+  constructor(private readonly userService: UserService) {}
+
+  @Get('/all')
+  @ApiOperation({
+    summary: 'Récupère tous les utilisateurs',
+  })
+  findAll() {
+    return this.userService.findAll();
+  }
+
+  @Get(':id')
+  @ApiOperation({
+    summary: 'Récupère un utilisateur grâce à son id',
+  })
+  findOne(@Param('id') id: string) {
+    return this.userService.findOne(id);
+  }
+
+  @Get('/')
+  @ApiOperation({
+    summary: 'Récupère un utilisateur grâce à son email',
+  })
+  async findOneByEmail(@Body('email') email: string) {
+    return this.userService.findOneByEmail(email);
+  }
+
+  @UseGuards(AuthGuard)
+  @Patch('/update')
+  @ApiOperation({
+    summary:
+      "Met à jour un utilisateur, seul l'utilisateur authentifié peut le faire",
+  })
+  update(@Req() request: Request, @Body() updateUserDto: UpdateUserDto) {
+    const id = request['user'].id;
+    return this.userService.update(id, updateUserDto);
+  }
+
+  @UseGuards(AuthGuard)
+  @Patch('/deactivate')
+  @ApiOperation({
+    summary:
+      "Désactive un utilisateur, seul l'utilisateur authentifié peut le faire",
+  })
+  deactivate(@Req() request: Request) {
+    const id = request['user'].id;
+    return this.userService.deactivate(id);
+  }
+
+  @UseGuards(AuthGuard)
+  @Delete('/delete')
+  @ApiOperation({
+    summary:
+      "Supprime un utilisateur, seul l'utilisateur authentifié peut le faire",
+  })
+  remove(@Req() request: Request) {
+    const id = request['user'].id;
+    return this.userService.remove(id);
+  }
 }