+++ CHANGELOG.md
@@ -7,3 +7,4 @@
 
 - suppression de la table Auth
 - la route /register renvoie l'access_token directement
+- update des routes user pour ne plus avoir à ajouter l'id en param dans les routes patch & delete